Refactor package names and create service layer.
