+++ src/main/java/com/sanitas/calculator/controller/CalculatorController.java
@@ -1,11 +1,9 @@
-package com.sanitas.calculator.controllers;
+package com.sanitas.calculator.controller;
 
 import com.sanitas.calculator.beans.OperationRequest;
 import com.sanitas.calculator.beans.OperationResponse;
 import com.sanitas.calculator.enums.OperationEnum;
-import com.sanitas.calculator.strategies.Calculation;
-import com.sanitas.calculator.strategies.SubtractOperation;
-import com.sanitas.calculator.strategies.SumOperation;
+import com.sanitas.calculator.service.CalculatorService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -19,6 +17,12 @@ import java.math.BigDecimal;
 @RequestMapping("/calculator")
 public class CalculatorController {
 
+  private final CalculatorService calculatorService;
+
+  CalculatorController(CalculatorService calculatorService) {
+    this.calculatorService = new CalculatorService();
+  }
+
   @PostMapping("/calculate")
   public ResponseEntity<OperationResponse> executeOperation(@RequestBody @Valid OperationRequest operationRequest) {
 
@@ -27,21 +31,11 @@ public class CalculatorController {
     final OperationEnum operation = operationRequest.getOperation();
     final OperationResponse operationResponse = new OperationResponse();
 
-    final BigDecimal result = defineAndCalculateOperation(number1, number2, operation);
+    final BigDecimal result = calculatorService.defineAndCalculateOperation(number1, number2, operation);
 
     operationResponse.setResult(result);
 
     return ResponseEntity.ok().body(operationResponse);
   }
 
-  private BigDecimal defineAndCalculateOperation(final BigDecimal number1, final BigDecimal number2,
-                                                        final OperationEnum operation) {
-    if (operation.equals(OperationEnum.SUBTRACTION)) {
-      final Calculation subtractionOperation = new Calculation(new SubtractOperation());
-      return subtractionOperation.execute(number1, number2);
-    } else {
-      final Calculation sumOperation = new Calculation(new SumOperation());
-      return sumOperation.execute(number1, number2);
-    }
-  }
 }
