+++ src/main/java/com/sanitas/calculator/validator/ValidOperationsValidator.java
@@ -1,4 +1,4 @@
-package com.sanitas.calculator.validators;
+package com.sanitas.calculator.validator;
 
 import com.sanitas.calculator.enums.OperationEnum;
