+++ src/test/java/com/sanitas/calculator/strategy/CalculationTest.java
@@ -1,4 +1,4 @@
-package com.sanitas.calculator.strategies;
+package com.sanitas.calculator.strategy;
 
 
 import org.junit.jupiter.api.Assertions;
